requirements.txt
httpx==0.25.0
attrs==23.1.0
cattrs==23.1.2
click==8.1.7
python-dateutil==2.8.2
structlog==23.2.0
xmltodict==0.13.0
pywebpush
python-dotenv==1.0.1
